+++ src/main/java/com/example/demo/service/AuthService.java
@@ -1,0 +1,7 @@
+package com.example.demo.service;
+
+import com.example.demo.dto.auth.CafegoryToken;
+
+public interface AuthService {
+	CafegoryToken refresh(String refreshToken);
+}
