+++ src/main/java/com/example/demo/service/OAuth2Service.java
@@ -5,6 +5,4 @@ import com.example.demo.dto.oauth2.OAuth2TokenRequest;
 
 public interface OAuth2Service {
 	CafegoryToken joinOrLogin(OAuth2TokenRequest oAuth2TokenRequest);
-
-	CafegoryToken refresh(String refreshToken);
 }
