+++ src/main/java/com/example/demo/service/OAuth2ServiceImpl.java
@@ -1,7 +1,5 @@
 package com.example.demo.service;
 
-import static com.example.demo.exception.ExceptionType.*;
-
 import java.util.Map;
 import java.util.Optional;
 
@@ -16,7 +14,6 @@ import com.example.demo.dto.auth.CafegoryToken;
 import com.example.demo.dto.oauth2.OAuth2Profile;
 import com.example.demo.dto.oauth2.OAuth2Token;
 import com.example.demo.dto.oauth2.OAuth2TokenRequest;
-import com.example.demo.exception.CafegoryException;
 import com.example.demo.repository.MemberRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -39,16 +36,6 @@ public class OAuth2ServiceImpl implements OAuth2Service {
 			return makeCafegoryToken(save.getId());
 		}
 		return makeCafegoryToken(byEmail.get().getId());
-	}
-
-	@Override
-	public CafegoryToken refresh(String refreshToken) {
-		boolean canRefresh = cafegoryTokenManager.canRefresh(refreshToken);
-		if (canRefresh) {
-			long identityId = cafegoryTokenManager.getIdentityId(refreshToken);
-			return makeCafegoryToken(identityId);
-		}
-		throw new CafegoryException(TOKEN_REFRESH_REJECT);
 	}
 
 	private CafegoryToken makeCafegoryToken(long memberId) {
